Java 11.0.2 migration (#32)
* Code compiles - execution NOT tested

* updating gitignore

* Removed intellij settings files

* Removed more intellij files

* Added exclusion for JDK classes.

* Fixed purchasing script for vendors that have listed coin types.

* Updated script to not kick off until the entire preload is complete.

* adds static name entry for Solo movie poster and tcg9 vendor entry

* clean up empty and orphaned object templates

* adds placeholder black market (static) spawns

* corrects entries for the video game table to correctly set it in tcg series 2 and remove series 1 console errors

* Updated gitignore and removed intellij project files

* Fixed appearance reference for thranta payroll and kashyyyk door, added skipLosCheck objvar due to cannit see issue. Requires updated src

* Fixed appearance and template for terminal (#2)

* Fixed appearance and template for terminal (#3)

* Fixed appearance and template for terminal (#4)

* Deleted another faulty/orphaned object template

* Fixed gcw ranks option on frog. Only issue is that it doesn't award the officer commands or badges.

* Fixed some unneeded java 11 changes

+++ sku.0/sys.server/compiled/game/script/conversation/loveday_romance_target.java
@@ -108,7 +108,7 @@ public class loveday_romance_target extends script.base_script
     {
         String file = "event/love_day";
         Vector randomOrder = utils.getResizeableIntArrayScriptVar(player, "traits_randomOrder");
-        String traitObjVar = "playingCupid.trait_0" + ((Integer)randomOrder.get(0)).intValue();
+        String traitObjVar = "playingCupid.trait_0" + (Integer) randomOrder.get(0);
         String message = getString(new string_id(file, "trait_" + getStringObjVar(npc, traitObjVar)));
         return message;
     }
@@ -116,7 +116,7 @@ public class loveday_romance_target extends script.base_script
     {
         String file = "event/love_day";
         Vector randomOrder = utils.getResizeableIntArrayScriptVar(player, "traits_randomOrder");
-        String traitObjVar = "playingCupid.trait_0" + ((Integer)randomOrder.get(4)).intValue();
+        String traitObjVar = "playingCupid.trait_0" + (Integer) randomOrder.get(4);
         String message = getString(new string_id(file, "trait_" + getStringObjVar(npc, traitObjVar)));
         return message;
     }
@@ -124,7 +124,7 @@ public class loveday_romance_target extends script.base_script
     {
         String file = "event/love_day";
         Vector randomOrder = utils.getResizeableIntArrayScriptVar(player, "traits_randomOrder");
-        String traitObjVar = "playingCupid.trait_0" + ((Integer)randomOrder.get(3)).intValue();
+        String traitObjVar = "playingCupid.trait_0" + (Integer) randomOrder.get(3);
         String message = getString(new string_id(file, "trait_" + getStringObjVar(npc, traitObjVar)));
         return message;
     }
@@ -132,7 +132,7 @@ public class loveday_romance_target extends script.base_script
     {
         String file = "event/love_day";
         Vector randomOrder = utils.getResizeableIntArrayScriptVar(player, "traits_randomOrder");
-        String traitObjVar = "playingCupid.trait_0" + ((Integer)randomOrder.get(2)).intValue();
+        String traitObjVar = "playingCupid.trait_0" + (Integer) randomOrder.get(2);
         String message = getString(new string_id(file, "trait_" + getStringObjVar(npc, traitObjVar)));
         return message;
     }
@@ -140,7 +140,7 @@ public class loveday_romance_target extends script.base_script
     {
         String file = "event/love_day";
         Vector randomOrder = utils.getResizeableIntArrayScriptVar(player, "traits_randomOrder");
-        String traitObjVar = "playingCupid.trait_0" + ((Integer)randomOrder.get(1)).intValue();
+        String traitObjVar = "playingCupid.trait_0" + (Integer) randomOrder.get(1);
         String message = getString(new string_id(file, "trait_" + getStringObjVar(npc, traitObjVar)));
         return message;
     }
